+++ spec/tcx_spec.rb
@@ -3,10 +3,10 @@
 require 'spec_helper'
 
 describe Tcx do
-  let(:tcx_files) { Dir.glob(File.join(File.dirname(__FILE__), 'data', '**', '*.tcx')) }
+  let(:tcx_files) { Dir.glob(File.join(File.dirname(__FILE__), 'data/**/*.tcx')) }
 
   describe '#load_file' do
-    Dir.glob(File.join(File.dirname(__FILE__), 'data', '**', '*.tcx')).each do |tcx_file|
+    Dir.glob(File.join(File.dirname(__FILE__), 'data/**/*.tcx')).each do |tcx_file|
       describe Pathname.new(tcx_file).relative_path_from(Pathname.new(File.dirname(__FILE__)).parent).to_s do
         let(:tcx) { described_class.load_file(tcx_file) }
 
@@ -32,7 +32,7 @@ describe Tcx do
   end
 
   describe '#load' do
-    let(:data) { File.read(File.join(File.dirname(__FILE__), 'data', 'running', 'multiple_running_activities.tcx')) }
+    let(:data) { File.read(File.join(File.dirname(__FILE__), 'data/running/multiple_running_activities.tcx')) }
     let(:tcx) { described_class.load(data) }
 
     it 'returns a database' do